sound: make pulseaudio optional again
