+++ Cargo.toml
@@ -12,6 +12,8 @@ authors = [
 edition = "2021"
 
 [features]
+default = ["pulseaudio"]
+pulseaudio = ["libpulse-binding"]
 debug_borders = [] # Make widgets' borders visible
 
 [dependencies]
@@ -23,8 +25,8 @@ env_logger = "0.9"
 hyper = "0.14"
 inotify = "0.10"
 libc = "0.2"
+libpulse-binding = { version = "2.0", default-features = false, optional = true }
 log = "0.4"
-libpulse-binding = { optional = true, version = "2.0", default-features = false }
 maildir = { version = "0.6", optional = true }
 neli = { version = "0.6", features = ["async"] }
 neli-wifi = { version = "0.3", features = ["async"] }
